pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtl"
version = "0.1.0"
description = "Git-Transform-Load: Sync Git repository history to BigQuery"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "GTL Contributors" }
]
keywords = ["git", "bigquery", "etl", "sync"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-cloud-bigquery>=3.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
gtl = "gtl.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gtl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
